+++ obsidian/plotting/plotly.py
@@ -5,13 +5,85 @@ from obsidian.optimizer import Optimizer
 from obsidian.exceptions import UnfitError, UnsupportedError
 from obsidian.parameters import Param_Continuous
 from .branding import obsidian_colors
+from obsidian.plotting.branding import obsidian_color_list as colors
 
 import plotly.graph_objects as go
 from plotly.graph_objects import Figure
+from plotly.subplots import make_subplots
 from sklearn.manifold import MDS
 
 import pandas as pd
 import numpy as np
+import math
+
+
+def visualize_inputs(campaign: Campaign) -> Figure:
+    """
+    Visualizes the input variables of a campaign.
+
+    Args:
+        campaign (Campaign): The campaign object containing the input data.
+
+    Returns:
+        Figure: The plotly Figure object containing the visualization.
+    """
+    n_dim = campaign.X_space.n_dim
+    X = campaign.X
+    
+    # Enforce that there are 2 rows
+    # Determine the number of columns based on the number of dimensions
+    rows = 2
+    cols = math.ceil(n_dim / rows)
+    
+    height = 200 * rows
+    width = 300 * cols
+    fontsize = 8
+
+    color_list = colors * 10
+    
+    # Add an extra 2 cols for the correlation matrix
+    fig = make_subplots(
+        rows=rows, cols=cols + 2,
+        vertical_spacing=0.2,
+        horizontal_spacing=0.1,
+        specs=[[{}]*cols + [{"colspan": 2, "rowspan": 2}, None],
+               [{}]*cols + [None, None]],
+        subplot_titles=[X.columns[i] for i in range(cols)]
+        + ['Correlation Matrix']
+        + [X.columns[i] for i in range(cols, n_dim)]
+    )
+    
+    for i, param in enumerate(X.columns):
+        row_i = i // cols + 1
+        col_i = i % cols + 1
+        fig.add_trace(go.Scatter(x=X.index, y=X[param],
+                                 mode='markers', name=param,
+                                 marker=dict(color=color_list[i]),
+                                 showlegend=False),
+                      row=row_i, col=col_i)
+        fig.update_xaxes(tickvals=np.around(np.linspace(0, campaign.m_exp, 5)),
+                         row=row_i, col=col_i)
+    
+    # Calculate the correlation matrix
+    X_u = campaign.X_space.unit_map(X)
+    corr_matrix = X_u.corr()
+    fig.add_trace(go.Heatmap(z=corr_matrix.values,
+                             x=corr_matrix.columns,
+                             y=corr_matrix.columns,
+                             colorscale=[[0, obsidian_colors.rich_blue],
+                                         [0.5, obsidian_colors.teal],
+                                         [1, obsidian_colors.lemon]],
+                             name='Correlation'),
+                  row=1, col=cols+1)
+    
+    fig.update_yaxes(showticklabels=False, row=1, col=cols+1)
+    fig.update_xaxes(tickangle=-90, row=1, col=cols+1)
+    
+    fig.update_layout(width=width, height=height, template='ggplot2',
+                      font_size=fontsize, title_text='Campaign Data Visualization')
+    fig.update_annotations(font_size=fontsize)
+    
+    return fig
 
 
 def MDS_plot(campaign: Campaign) -> Figure: