+++ obsidian/campaign/campaign.py
@@ -241,6 +241,19 @@ class Campaign():
             return self.y
     
     @property
+    def X_best(self) -> pd.DataFrame:
+        """
+        Best performing X values
+        """
+        best_idx = self.out.idxmax().values
+        
+        X_best = self.X.iloc[best_idx, :]
+        if isinstance(X_best, pd.Series):
+            X_best = X_best .to_frame().T
+        
+        return X_best
+    
+    @property
     def X(self) -> pd.DataFrame:
         """
         Feature columns of the training data