+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Modified
 - Switched all usages of X_ref = X_space.mean() to optimizer.X_best_f
+- Refactored mpl "visualize_inputs" as plotly "visualize_inputs" for better interactivity
 
 ## [0.8.3]
 ### Added