+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.8.4]
+### Added
+- Campaign X_best method
+- Optimizer X_best_f attribute(s)
+- Sequence of colors "color_list" to branding
+
+### Modified
+- Switched all usages of X_ref = X_space.mean() to optimizer.X_best_f
+- Refactored mpl "visualize_inputs" as plotly "visualize_inputs" for better interactivity
+
 ## [0.8.3]
 ### Added
 - Default values for NParEGO scalarization_weights